requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
gitpython==3.1.40
pytest==7.4.3
pytest-cov==4.1.0
pytest-timeout==2.2.0
astroid==3.0.1
pylint==3.0.2
flake8==6.1.0
requests==2.31.0
httpx==0.25.1
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
websockets==12.0
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
chromadb==0.4.18
sentence-transformers==2.2.2
slack-sdk==3.26.1
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.7.0
asyncio==3.4.3
aiohttp==3.9.1
watchdog==3.0.0